+++ executor/loadremotetest/one_csv_test.go
@@ -19,7 +19,6 @@ import (
 
 	"github.com/fsouza/fake-gcs-server/fakestorage"
 	"github.com/pingcap/tidb/testkit"
-	"github.com/stretchr/testify/require"
 )
 
 func (s *mockGCSSuite) TestLoadCSV() {
@@ -266,194 +265,4 @@ mynull,"mynull"
 		FIELDS TERMINATED BY ',' DEFINED NULL BY 'mynull' OPTIONALLY ENCLOSED
 		LINES TERMINATED BY '\n';`, gcsEndpoint)
 	s.tk.MustMatchErrMsg(sql, `must specify FIELDS \[OPTIONALLY\] ENCLOSED BY`)
-}
-
-func (s *mockGCSSuite) TestGBK() {
-	s.tk.MustExec("DROP DATABASE IF EXISTS load_charset;")
-	s.tk.MustExec("CREATE DATABASE load_charset;")
-	s.tk.MustExec(`CREATE TABLE load_charset.gbk (
-    	i INT, j VARCHAR(255)
-    	) CHARACTER SET gbk;`)
-	s.tk.MustExec(`CREATE TABLE load_charset.utf8mb4 (
-    	i INT, j VARCHAR(255)
-    	) CHARACTER SET utf8mb4;`)
-
-	s.server.CreateObject(fakestorage.Object{
-		ObjectAttrs: fakestorage.ObjectAttrs{
-			BucketName: "test-load",
-			Name:       "gbk.tsv",
-		},
-		Content: []byte{
-			// 1	一丁丂七丄丅丆万丈三上下丌不与丏
-			0x31, 0x09, 0xd2, 0xbb, 0xb6, 0xa1, 0x81, 0x40, 0xc6, 0xdf, 0x81,
-			0x41, 0x81, 0x42, 0x81, 0x43, 0xcd, 0xf2, 0xd5, 0xc9, 0xc8, 0xfd,
-			0xc9, 0xcf, 0xcf, 0xc2, 0xd8, 0xa2, 0xb2, 0xbb, 0xd3, 0xeb, 0x81,
-			0x44, 0x0a,
-			// 2	丐丑丒专且丕世丗丘丙业丛东丝丞丢
-			0x32, 0x09, 0xd8, 0xa4, 0xb3, 0xf3, 0x81, 0x45, 0xd7, 0xa8, 0xc7,
-			0xd2, 0xd8, 0xa7, 0xca, 0xc0, 0x81, 0x46, 0xc7, 0xf0, 0xb1, 0xfb,
-			0xd2, 0xb5, 0xb4, 0xd4, 0xb6, 0xab, 0xcb, 0xbf, 0xd8, 0xa9, 0xb6,
-			0xaa,
-		},
-	})
-
-	sql := fmt.Sprintf(`LOAD DATA INFILE 'gs://test-load/gbk.tsv?endpoint=%s'
-		INTO TABLE load_charset.gbk CHARACTER SET gbk`, gcsEndpoint)
-	s.tk.MustExec(sql)
-	s.tk.MustQuery("SELECT * FROM load_charset.gbk;").Check(testkit.Rows(
-		"1 一丁丂七丄丅丆万丈三上下丌不与丏",
-		"2 丐丑丒专且丕世丗丘丙业丛东丝丞丢",
-	))
-	sql = fmt.Sprintf(`LOAD DATA INFILE 'gs://test-load/gbk.tsv?endpoint=%s'
-		INTO TABLE load_charset.utf8mb4 CHARACTER SET gbk`, gcsEndpoint)
-	s.tk.MustExec(sql)
-	s.tk.MustQuery("SELECT * FROM load_charset.utf8mb4;").Check(testkit.Rows(
-		"1 一丁丂七丄丅丆万丈三上下丌不与丏",
-		"2 丐丑丒专且丕世丗丘丙业丛东丝丞丢",
-	))
-
-	s.tk.MustExec("TRUNCATE TABLE load_charset.utf8mb4;")
-	s.tk.MustExec("SET SESSION character_set_database = 'gbk';")
-	sql = fmt.Sprintf(`LOAD DATA INFILE 'gs://test-load/gbk.tsv?endpoint=%s'
-		INTO TABLE load_charset.utf8mb4;`, gcsEndpoint)
-	s.tk.MustExec(sql)
-	s.tk.MustQuery("SELECT * FROM load_charset.utf8mb4;").Check(testkit.Rows(
-		"1 一丁丂七丄丅丆万丈三上下丌不与丏",
-		"2 丐丑丒专且丕世丗丘丙业丛东丝丞丢",
-	))
-
-	s.server.CreateObject(fakestorage.Object{
-		ObjectAttrs: fakestorage.ObjectAttrs{
-			BucketName: "test-load",
-			Name:       "utf8mb4.tsv",
-		},
-		Content: []byte("1\t一丁丂七丄丅丆万丈三上下丌不与丏\n" +
-			"2\t丐丑丒专且丕世丗丘丙业丛东丝丞丢"),
-	})
-
-	s.tk.MustExec("TRUNCATE TABLE load_charset.gbk;")
-	sql = fmt.Sprintf(`LOAD DATA INFILE 'gs://test-load/utf8mb4.tsv?endpoint=%s'
-		INTO TABLE load_charset.gbk CHARACTER SET utf8mb4`, gcsEndpoint)
-	s.tk.MustExec(sql)
-	s.tk.MustQuery("SELECT * FROM load_charset.gbk;").Check(testkit.Rows(
-		"1 一丁丂七丄丅丆万丈三上下丌不与丏",
-		"2 丐丑丒专且丕世丗丘丙业丛东丝丞丢",
-	))
-
-	s.server.CreateObject(fakestorage.Object{
-		ObjectAttrs: fakestorage.ObjectAttrs{
-			BucketName: "test-load",
-			Name:       "emoji.tsv",
-		},
-		Content: []byte("1\t一丁丂七😀😁😂😃\n" +
-			"2\t丐丑丒专😄😅😆😇"),
-	})
-
-	s.tk.MustExec("TRUNCATE TABLE load_charset.gbk;")
-	sql = fmt.Sprintf(`LOAD DATA INFILE 'gs://test-load/emoji.tsv?endpoint=%s'
-		INTO TABLE load_charset.gbk CHARACTER SET utf8mb4`, gcsEndpoint)
-	err := s.tk.ExecToErr(sql)
-	checkClientErrorMessage(s.T(), err, `ERROR 1366 (HY000): Incorrect string value '\xF0\x9F\x98\x80' for column 'j'`)
-
-	sql = fmt.Sprintf(`LOAD DATA INFILE 'gs://test-load/emoji.tsv?endpoint=%s'
-		IGNORE INTO TABLE load_charset.gbk CHARACTER SET utf8mb4`, gcsEndpoint)
-	s.tk.MustExec(sql)
-	require.Equal(s.T(), "Records: 2  Deleted: 0  Skipped: 0  Warnings: 2", s.tk.Session().GetSessionVars().StmtCtx.GetMessage())
-	s.tk.MustQuery("SELECT HEX(j) FROM load_charset.gbk;").Check(testkit.Rows(
-		"D2BBB6A18140C6DF3F3F3F3F",
-		"D8A4B3F38145D7A83F3F3F3F",
-	))
-
-	sql = fmt.Sprintf(`LOAD DATA INFILE 'gs://test-load/gbk.tsv?endpoint=%s'
-		INTO TABLE load_charset.utf8mb4 CHARACTER SET unknown`, gcsEndpoint)
-	err = s.tk.ExecToErr(sql)
-	require.ErrorContains(s.T(), err, "Unknown character set: 'unknown'")
-}
-
-func (s *mockGCSSuite) TestOtherCharset() {
-	s.tk.MustExec("DROP DATABASE IF EXISTS load_charset;")
-	s.tk.MustExec("CREATE DATABASE load_charset;")
-	s.tk.MustExec(`CREATE TABLE load_charset.utf8 (
-    	i INT, j VARCHAR(255)
-    	) CHARACTER SET utf8;`)
-	s.tk.MustExec(`CREATE TABLE load_charset.utf8mb4 (
-    	i INT, j VARCHAR(255)
-    	) CHARACTER SET utf8mb4;`)
-
-	s.server.CreateObject(fakestorage.Object{
-		ObjectAttrs: fakestorage.ObjectAttrs{
-			BucketName: "test-load",
-			Name:       "utf8.tsv",
-		},
-		Content: []byte("1\tကခဂဃ\n2\tငစဆဇ"),
-	})
-
-	sql := fmt.Sprintf(`LOAD DATA INFILE 'gs://test-load/utf8.tsv?endpoint=%s'
-		INTO TABLE load_charset.utf8 CHARACTER SET utf8`, gcsEndpoint)
-	s.tk.MustExec(sql)
-	s.tk.MustQuery("SELECT * FROM load_charset.utf8;").Check(testkit.Rows(
-		"1 ကခဂဃ",
-		"2 ငစဆဇ",
-	))
-	sql = fmt.Sprintf(`LOAD DATA INFILE 'gs://test-load/utf8.tsv?endpoint=%s'
-		INTO TABLE load_charset.utf8mb4 CHARACTER SET utf8`, gcsEndpoint)
-	s.tk.MustExec(sql)
-	s.tk.MustQuery("SELECT * FROM load_charset.utf8mb4;").Check(testkit.Rows(
-		"1 ကခဂဃ",
-		"2 ငစဆဇ",
-	))
-
-	s.server.CreateObject(fakestorage.Object{
-		ObjectAttrs: fakestorage.ObjectAttrs{
-			BucketName: "test-load",
-			Name:       "latin1.tsv",
-		},
-		// "1\t‘’“”\n2\t¡¢£¤"
-		Content: []byte{0x31, 0x09, 0x91, 0x92, 0x93, 0x94, 0x0a, 0x32, 0x09, 0xa1, 0xa2, 0xa3, 0xa4},
-	})
-	s.tk.MustExec(`CREATE TABLE load_charset.latin1 (
-    	i INT, j VARCHAR(255)
-    	) CHARACTER SET latin1;`)
-	sql = fmt.Sprintf(`LOAD DATA INFILE 'gs://test-load/latin1.tsv?endpoint=%s'
-		INTO TABLE load_charset.latin1 CHARACTER SET latin1`, gcsEndpoint)
-	s.tk.MustExec(sql)
-	s.tk.MustQuery("SELECT * FROM load_charset.latin1;").Check(testkit.Rows(
-		"1 ‘’“”",
-		"2 ¡¢£¤",
-	))
-
-	s.tk.MustExec("TRUNCATE TABLE load_charset.utf8mb4;")
-	sql = fmt.Sprintf(`LOAD DATA INFILE 'gs://test-load/latin1.tsv?endpoint=%s'
-		INTO TABLE load_charset.utf8mb4 CHARACTER SET latin1`, gcsEndpoint)
-	s.tk.MustExec(sql)
-	s.tk.MustQuery("SELECT * FROM load_charset.utf8mb4;").Check(testkit.Rows(
-		"1 ‘’“”",
-		"2 ¡¢£¤",
-	))
-
-	s.server.CreateObject(fakestorage.Object{
-		ObjectAttrs: fakestorage.ObjectAttrs{
-			BucketName: "test-load",
-			Name:       "ascii.tsv",
-		},
-		Content: []byte{0, 1, 2, 3, 4, 5, 6, 7},
-	})
-	s.tk.MustExec(`CREATE TABLE load_charset.ascii (
-    	j VARCHAR(255)
-    	) CHARACTER SET ascii;`)
-	s.tk.MustExec(`CREATE TABLE load_charset.binary (
-    	j VARCHAR(255)
-    	) CHARACTER SET binary;`)
-	sql = fmt.Sprintf(`LOAD DATA INFILE 'gs://test-load/ascii.tsv?endpoint=%s'
-		INTO TABLE load_charset.ascii CHARACTER SET ascii`, gcsEndpoint)
-	s.tk.MustExec(sql)
-	s.tk.MustQuery("SELECT HEX(j) FROM load_charset.ascii;").Check(testkit.Rows(
-		"0001020304050607",
-	))
-	sql = fmt.Sprintf(`LOAD DATA INFILE 'gs://test-load/ascii.tsv?endpoint=%s'
-		INTO TABLE load_charset.binary CHARACTER SET binary`, gcsEndpoint)
-	s.tk.MustExec(sql)
-	s.tk.MustQuery("SELECT HEX(j) FROM load_charset.binary;").Check(testkit.Rows(
-		"0001020304050607",
-	))
 }
